+++ src/app.ts
@@ -7,7 +7,7 @@ import { getAllJobRoles } from "./controllers/JobRoleController";
 import { UserRole } from "./models/JwtToken";
 import { allowRoles, setLoggedInStatus } from "./middleware/AuthMiddleware";
 import { v4 as uuidv4 } from 'uuid';
-import { getChatGPTResponse } from "../src/utils/interviewBot";
+import { getChatGPTResponse } from "./services/InterviewService";
 import { getApplicationForm, postApplicationForm } from "./controllers/ApplicationController";
 
 const app = express();
@@ -66,7 +66,7 @@ app.get('/job-roles', allowRoles([UserRole.Admin, UserRole.User]),getAllJobRoles
 app.get('/applicationForm', getApplicationForm)
 app.post('/applicationForm', postApplicationForm)
 
-// AI Mock Interview Assistan
+// AI Mock Interview Assistant
 const OPENAI_API_KEY = process.env.INTERVIEW_AI_API_KEY;
 
 interface InterviewSession {
