+++ views/interviewChat.html
@@ -75,6 +75,6 @@
       <button id="chat-button" class="chat-button">Send</button>
     </div>
   </div>
-  <script src="../dist/index.js"></script>
+  <script src="../dist/client/InterviewBotService.js"></script>
 </body>
 </html>
